Move getIndicesWithRegex to different directory
