+++ src/researches/passivevoice/getIndicesWithRegex.js
